[libc][math] Refactor exp10f implementation to header-only in src/__support/math folder.
