+++ libc/shared/math.h
@@ -13,6 +13,7 @@
 
 #include "math/exp.h"
 #include "math/exp10.h"
+#include "math/exp10f.h"
 #include "math/expf.h"
 #include "math/expf16.h"
 #include "math/frexpf.h"
