+++ libc/shared/math/exp10f.h
@@ -1,4 +1,5 @@
-//===-- Single-precision 10^x function ------------------------------------===//
+//===-- Shared exp10f function -----------------------------------*- C++
+//-*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,12 +7,18 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "src/math/exp10f.h"
+#ifndef LLVM_LIBC_SHARED_MATH_EXP10F_H
+#define LLVM_LIBC_SHARED_MATH_EXP10F_H
 
+#include "shared/libc_common.h"
 #include "src/__support/math/exp10f.h"
 
 namespace LIBC_NAMESPACE_DECL {
+namespace shared {
 
-LLVM_LIBC_FUNCTION(float, exp10f, (float x)) { return math::exp10f(x); }
+using math::exp10f;
 
+} // namespace shared
 } // namespace LIBC_NAMESPACE_DECL
+
+#endif // LLVM_LIBC_SHARED_MATH_EXP10F_H
